requirements.txt
torch
torchvision
timm
pandas
Pillow
opencv-python
scipy
numpy
wandb
efficientkan
